Create README.md
In this example, using Visual Studio 2022 we will create a Client-Server example of a customer service that gets the customer data and handle it internally like for example saving it to the database “I will not implement this part because it is out of the article scope”, and the client that will send the data to the service and get response according to it. After that, we will create a docker image for the service and then deploy it to Kubernetes.